[FIX] account_reports: fix overlapping thead
It seems that wkhtmltopdf with patched qt have an issue of overlapping a
duplicated thead when a page break happen inside a row directly
following the thead.

The solution given on wkhtmltopdf:
 https://github.com/wkhtmltopdf/wkhtmltopdf/issues/1524

is to use the CSS `tr { page-break-inside: avoid }`, this is already
used in Odoo in multiple locations (eg. odoo/odoo@10e4a98b8) but was
not present in tables of financial reports.

opw-743033
closes #1018