[FIX] web_enterprise: make the loading message always visible
The "Loading..." message which is showing once an action is triggered in
the backend was sometimes hidden by other content such as kanban record
or even the appswitcher itself. As it is positioned in the <body/>
element, setting its z-index to be above everything should be enough to
fix the problem.

opw-743458