[FIX] mrp_account: set default value to avoid crash on report
When clicking on "Cost Analysis" on product form view,
a html report should open, if all Manufacturing order
for that product are in 'done' state.
The report rendering crashes if not value for 'Cost
per hour' field (costs_hour) is set for Workcenter
of related MO. Indeed, the qweb report try to convert
that NULL costs into current user currency (NoneTyep
can not be divised).
The fix consist by setting a default value for the hour
cost.

opw-743708