[FIX] inter_company_rules: get the other company's view point
commit 45451b2c6e7e8070b9d77ed2ad92bce4d37bf5f9 introduced a bug while solving another one.
First bug opw 727477: When in A we do an PO on B, a SO will be created in B
the SO must take A as its customer, as viewed from B (to get the pricelist right)
BUT we are currently in A.
HENCE we need to partner_A.sudo(B_intercompany_user_uid) to make it right.

Second bug that the fix has brought:
If the intercompany_users are not part of the other company as well,
we end up trying to access A's PO with userB's rights, which raises an error.

In the context of sharing partners between companies, we can safely go
in A from the company of the PO to the partner representing A, but viewed from B
in order to make a SO

This also works when partners are not shared BUT both A's and B's partners has company _id = Null

OPW 743900
and
OPW 727477

Closes #1035