[FIX] website_sign: prevent crash on main kanban in mobile mode
A customization is made for the kanban view of the 'signature.request'
model so that the records cannot be dragged. This was done by destroying
the jquery.ui 'sortable' instance at the end of the `start` method. The
problem is that this instance was not created in all cases (e.g. mobile
mode) and trying to destroy it was making the webclient crash.

Hopefully some improvement has been made by commit:
https://github.com/odoo/odoo/commit/7a9c16ffee3e1154f78cfa04719d752f74fa8a77
It indeed adds the possibility to disable the draggable feature thanks
to an option, which makes the 'website_sign' customization simpler:
forcing the option to false in the `init` method for the 'signature.request'
model.

opw-744282