[FIX] web_studio: prevent crash when evaluating dynamic contexts
Studio uses a pre-version of the new views (merged in saas-16)
and this implementation is experimental/incompte/incorrect
(sometimes)...

In studio, dynamic contexts can't be evaluated because the evaluation
context isn't passed, and can't even be computed correclty without
large development.

Now, we prevent the crash by protecting the evaluation with a try/catch,
so that static are still passed correclty. This may produce slight
rendering differences between the rendering in and outside Studio.

opw-708481
opw-743761