[FIX] web_studio: wrong get_value for default_get
In the case of a 'read' RPC, the value of many2one fields is an
array [id, display_name], whereas in the case of a 'default_get',
the value is the id.
The former case wasn't correctly handled by the model.

It caused a crash when trying to evaluate a domain when opening
the form view of crm.lead in Studio, with no record.

opw-743761