[FIX] account_taxcloud: Update correctly taxes on SO and INV validations
What are the steps to reproduce your issue?

1. Visit RUNBOT for Enterprise Version
2. Accounting --> Configuration.  Enter TaxCloud API ID. Enter TaxCloud API KEY.  Click SYNC TAXCLOUD CATEGORY (TIC).  Click APPLY
3. Create a new Product called "TEST" (just give it a name, accept all defaults for other fields)
4. Create "CUSTOMER 1" with address: 4210 E Highland Ave, Highland, CA 92346 and Fiscal Position (Accounting Tab) "Fiscal Position for United States".
5. Create "CUSTOMER 2" with address: 7201 Boulder Avenue, Highland, CA, 92346 and Fiscal Position (Accounting Tab) "Fiscal Position for United States".

What is the current behavior that you observe?

6. Create a new Sales Order and sell TEST to CUSTOMER 1 - you should see a Tax Rate of 8% - confirm the order.  The tax rate is unchanged.
7. Create a new Sales Order and sell TEST to CUSTOMER 2 - you should see a Tax Rate of 8% - confirm the order.  The tax rate is unchanged.

What would be your expected behavior in this case?

The second rate is incorrect.

If you delete the TAX MAPPING that is stored after the first Sales Order in the Fiscal Position:
a)  Accounting --> Configuration --> Fiscal Positions.
b) Open "Fiscal Position for United States"
c) Click EDIT, delete the Tax Mapping line, click SAVE

And then create a new SO to CUSTOMER 2, the tax rate this time is 7.75%.  Now create a new SO to CUSTOMER 1, the tax rate this time is 7.75%.