[FIX] web_studio: deal with unrelated ´ir.model.data´ during export
During the Studio export, we use all ´ir.model.data´ created by Studio.
However, it may happen that an ´ir.model.data´ is linked to an inexisting record in database.
In this case, the export will crash with an error.

This commit handles this case by ignoring these broken ´ir.model.data´.