[FIX] account_yodlee: reinstallation works now
When an user tries to uninstall then reinstall the Yodlee module (for
some reasons), the installation process crashes.

This is due to the fact that Yodlee's online.institution records are
inserted in an non-ORM way (in SQL), and thus not removed when
uninstalling.

This fix proposes to mute the error when the module tries to insert an
online.institution that is still in the database.

(We could also have a post uninstall hook that removes them all, but we
don't like removing data in our users databases.)