[FIX] account_reports: general ledger cash basis with analytic tags/accounts
In the General ledger report,
when selecting the cash basis method and then an analytic
account or tags,
the sql requests failed because of a missing column
`analytic_account_id`.

This ie because it wasn't set in the with statement of the
given sql request. This is now the case with this revision.

opw-745691