[FIX] account_reports: ensure that a followup context exists for uid
The /account_reports/followup_report/ route will search for contexts
belonging to a partner and the current user. Without this the
following scenario returns empty pdf's because the followup context is
never created for the current user:

1. demo user opens customers statement
2. admin opens customers statement and clicks on Print Letter

Note that this problem automatically disappears after a while because
account.report.context.followup.all is a TransientModel whose records
will be deleted daily by the 'Auto-vacuum internal data' cron.

opw-746370