[FIX] web_gantt: hide records without name value
If a record has no name (e.g. on hr.holidays it is not required), the gantt used
to display 'false' as the name of the record.

In saas-15, a more generic solution of using display_name instead of name has
been done at afd68a7d.
This patch is a safer fix for stable version.

Closes odoo/odoo#17451