[FIX] web_studio: avoid to reuse the same xpath node
Before this fix, some nodes were reuse to simplify the arch view.
If a xpath was already present in the view, the operation was
then added to that xpath node. If not, a new xpath node was created
at the end of the arch view.

The problem is that sometimes we need to create absolute paths like:
/form[1]/sheet[1]/group[2]/group[2].
So, if another operation is added before this type of xpath,
this absolute xpath will probably no longer target the right element.

Steps to reproduce (sale.order form):
- add an element after the title
- add columns before the notebook
- add columns after the title
The first columns will not be placed at the right place.

Now, we decided to remove this optimization.
Each new operations will be added to the end of the arch.

opw-745364