[FIX] mrp_workorder: traceability report
In case of products tracked by serial number or lots, duplicated quants
can appear when performing unbuilding. This is confusing for the end
user, since he will see an incorrect number of quants. For example, in
the case of a unique serial number, two quants with the same serial
number will be shown.

We cannot avoid this situation with the current architecture. However,
we can hide the duplicated quant in the traceability report.

We do it for products tracked by unique serial number only, by assuming
that the quant with the highest id is the correct one. This should be a
fair assumption, as a strict check would require the computation of the
complete history of the quant, which would be prohibitive for the added
value.

opw-743874